Bugfixes for #2, notably reordering post-ability steps
Summary:

Especially at level 80, we'd occasionally generate a sequence that would be
unsuccessful even though the simulation thought it was 100% success rate.
There were two causes for this:
1. the resolution order of post-ability actions was incorrect. This was
especially a problem with manipulation, where we'd resolve the effect before
the durability decrease, leading to us thinking we'd never hit 0 durability
when we actually had.
2. There are lots of high-quality ability series that use rapid synthesis +
steady hand II to generate large jump in progress at 90% probability. Because
we only simulate 20 times per ability chain, this means that sometimes we'd
never hit a failure, and think it was a 100% success rate chain.

These are fixed in this commit by:
1. fixing the resolution order so that post-action effects are resolved after
the durability decrease
2. always doing a worst-case sim and including that in the success rate for
scoring.

Fixes #2