Add some heuristics for difficult crafts
If a craft is very difficult to get to succeed (i.e. it requires many
progress steps), it's possible we won't happen upon a random sequence that
succeeds. In that case, we'd just choose the one with highest quality, but
that's not really useful.

This adds two additional changes to help find a successful solution in these
cases:
- if the success rate is 0, we default to swapping a quality for a progress
  ability.
- we score non-100%-success scaled by the successRate. This lets us give some
  additional weight to higher success rates and helps us move continuously
  towards 100% success, rather than having a discontinuity