Fix a rounding error in progress / quality simulation
We ought to have been rounding down rather than rounding to the nearest
integer on progress / quality increases. This was not really a problem for
quality, but for progress it could lead to us being just short of the
difficulty and failing the craft when we thought it would be successful.

Fixes #3