Add a heuristic to remove unused no-op abilities
It often happens that we get no-op abilities (like observe without a combo
ability) stuck in the series of steps because there's no strong pressure to
remove them, and it's rare that we'll randomly try to take them out.

This commit adds a heuristic that tries removing them at the end of each
generation and accepts/rejects according to normal criteria.